+++ src/pages/add-doctor/add-doctor.html
@@ -42,12 +42,12 @@
 
     <ion-item>
       <ion-label>Working hours (start)</ion-label>
-      <ion-datetime displayFormat="HH:mm" pickerFormat="HH mm" [(ngModel)]="defaultTime.timeStarts"></ion-datetime>
+      <ion-datetime displayFormat="HH:mm" pickerFormat="HH mm" [(ngModel)]="doctor.workHoursStart"></ion-datetime>
     </ion-item>
 
     <ion-item>
       <ion-label>Working hours (end)</ion-label>
-      <ion-datetime displayFormat="HH:mm" pickerFormat="HH mm" [(ngModel)]="defaultTime.timeEnds"></ion-datetime>
+      <ion-datetime displayFormat="HH:mm" pickerFormat="HH mm" [(ngModel)]="doctor.workHoursEnd"></ion-datetime>
     </ion-item>
 
     <ion-item>
