+++ src/pages/add-doctor/add-doctor.ts
@@ -5,6 +5,8 @@ import { Speciality } from '../../models/speciality';
 import { Room } from '../../models/room';
 import { Observable } from 'rxjs/Rx';
 import { DoctorProvider } from '../../providers/doctor/doctor';
+import { RoomProvider } from '../../providers/room/room';
+import { SpecialityProvider } from '../../providers/speciality/speciality';
 
 @Component({
   selector: 'page-doctor',
@@ -13,23 +15,19 @@ import { DoctorProvider } from '../../providers/doctor/doctor';
 
 export class DoctorRegistration {
   public doctor = {};
-  public defaultTime = {
-    month: '2018-01-01',
-    timeStarts: '08:00',
-    timeEnds: '17:00',
-  };
   private doctorObservable: Observable<Doctor[]>;
   private roomObservable: Observable<Room[]>;
   private specialityObservable: Observable<Speciality[]>;
 
-  constructor(public navCtrl: NavController, public alertCtrl: AlertController, public doctorProvider: DoctorProvider) {
+  constructor(public navCtrl: NavController, public alertCtrl: AlertController, public doctorProvider: DoctorProvider, public roomProvider: RoomProvider, specialityProvider: SpecialityProvider) {
     this.doctorObservable = doctorProvider.getDoctor();
-    this.roomObservable = doctorProvider.getRooms();
-    this.specialityObservable = doctorProvider.getSpecialities();
+    this.roomObservable = roomProvider.getRooms();
+    this.specialityObservable = specialityProvider.getSpecialities();
   }
 
   public addDoctor(doctor: Doctor) {
     this.doctorProvider.addDoctor(doctor);
+    this.navCtrl.pop();
   }
 
   public goBack() {
